# internal/handlers/story/create_result.go
package handlers

import (
	"net/http"
	db "stories-backend/pkg/db/mongo"

	domain "stories-backend/internal/domain/story"
	handlers "stories-backend/internal/handlers/common"

	"github.com/gin-gonic/gin"
)

// @Summary Create or update story result
// @Description Creates or updates a user's result for a specific story scene
// @Tags Story
// @Accept json
// @Produce json
// @Param request body createResultBody true "Result data"
// @Success 201 {object} domain.GetStoryResult "Result created/updated successfully"
// @Failure 400 {object} handlers.BaseErrorResponse "Invalid request body or ID format"
// @Failure 401 {object} handlers.BaseErrorResponse "Unauthorized"
// @Failure 500 {object} handlers.BaseErrorResponse "Internal server error"
// @Security ApiKeyAuth
// @Router /stories/{id}/results [put]
func (handler *StoryHandler) CreateResult(ctx *gin.Context) {
	var body createResultBody
	if err := ctx.ShouldBindJSON(&body); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "invalid request body"})
		return
	}

	userID, err := handlers.GetUserID(ctx)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	storyID, err := db.ParseObjectID(body.StoryID)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	sceneID, err := db.ParseObjectID(body.SceneID)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	result, err := handler.service.SetResult(domain.SetResultDTO{
		UserID:  userID,
		StoryID: storyID,
		SceneID: sceneID,
	})
	if err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	ctx.JSON(http.StatusCreated, result)
}

type createResultBody struct {
	StoryID string `json:"storyId"`
	SceneID string `json:"sceneId"`
}